+++ examples/src/FONTS_GLYPH/main.cpp
@@ -35,7 +35,7 @@ ERM19264_UC1609 myLCD(myLCDwidth ,myLCDheight, RST, CD );
 #define MY_TEST_DELAY1 bcm2835_delay(1000)
 
 // =============== Function prototype ================
-void setup(void);
+bool setup(void);
 void myTest(void);
 void EndTest(void);
 void testReset(void);
@@ -43,9 +43,7 @@ void testReset(void);
 // ======================= Main ===================
 int main(int argc, char **argv)
 {
-	if(!bcm2835_init()) {return -1;}
-
-	setup();
+	if(!setup()) {return -1;}
 	myTest();
 	EndTest();
 
@@ -55,18 +53,34 @@ int main(int argc, char **argv)
 
 
 // ===================== Function Space =====================
-void setup() {
-	bcm2835_delay(50);
-	printf("LCD Begin\r\n");
-	myLCD.LCDbegin(RAMaddressCtrl, LCDcontrast, SPICLK_FREQ , SPI_CE_PIN); // initialize the LCD
-	myLCD.LCDFillScreen(0x33); // splash screen bars
+bool setup() {
+	printf("LCD Test Begin\r\n");
+	// Check if Bcm28235 lib installed and print version.
+	if(!bcm2835_init())
+	{
+		printf("Error 1201: init bcm2835 library , Is it installed ?\r\n");
+		return false;
+	}else
+	{
+		printf("bcm2835 library Version Number :: %u\r\n",bcm2835_version());
+		bcm2835_delay(100);
+	}
+
+	if(!myLCD.LCDbegin(RAMaddressCtrl, LCDcontrast, SPICLK_FREQ , SPI_CE_PIN))  // initialize the LCD
+	{
+		printf("Error 1202: bcm2835_spi_begin :Cannot start spi, Running as root?\n");
+		return false;
+	}
+	printf("ERM19264 Library version number :: %u \r\n", myLCD.LCDLibVerNumGet());
+	myLCD.LCDFillScreen(0x33); // display splash screen bars, optional for effect
 	bcm2835_delay(1500);
+	return true;
 }
 
 void EndTest()
 {
 	myLCD.LCDPowerDown();
-	bcm2835_close(); // Close the library
+	bcm2835_close(); // Close library, deallocating allocated memory & closing /dev/mem
 	printf("LCD End\r\n");
 }
 
@@ -75,6 +89,12 @@ void myTest() {
 	// define a buffer to cover whole screen
 	uint8_t screenBuffer[myLCDwidth * (myLCDheight/8)]; 
 	myLCD.LCDbufferScreen = (uint8_t*) &screenBuffer;
+	if(myLCD.LCDbufferScreen== nullptr) // check if pointer is still = null
+	{
+		printf("Error 1203 :: Problem assigning buffer pointer\r\n");
+		exit(-1);
+	}
+	
 	myLCD.LCDclearBuffer();   // Clear buffer
 
 	// Test 1 Font FreeSans12pt7b
@@ -97,7 +117,7 @@ void myTest() {
 	myLCD.drawCharGlyph(10, 62, '!', FOREGROUND, 1);
 	testReset();
 
-	// Test 4 Font FreeSErif18pt7b
+	// Test 4 Font FreeSerif18pt7b
 	txt = "Free Serif     18pt7b";
 	myLCD.setFontGlyph(&FreeSerif18pt7b);
 	myLCD.drawTextGlyph(0, 26, txt, FOREGROUND, 1);