+++ examples/src/TEXT_GRAPHICS/main.cpp
@@ -30,8 +30,8 @@ ERM19264_UC1609 myLCD(myLCDwidth ,myLCDheight , RST, CD) ;
 #define PI 3.14
 
 // =============== Function prototype ================
-void setup(void);
-void myTest(void);
+bool setup(void);
+bool myTest(void);
 void DisplayText();
 void DisplayGraphics();
 void EndTest(void);
@@ -40,32 +40,43 @@ void TestReset(void);
 // ======================= Main ===================
 int main(int argc, char **argv)
 {
-	if(!bcm2835_init()){return -1;}
-
-	setup();
-	myTest();
+	if(!setup()) {return -1;}
+	if(!myTest()){return -1;}
 	EndTest();
 
 	return 0;
 }
 // ======================= End of main  ===================
 
-
-
 // ===================== Function Space =====================
-void setup() {
-	bcm2835_delay(50);
-	printf("LCD Begin\r\n");
-	myLCD.LCDbegin(RAMaddressCtrl, LCDcontrast, SPICLK_FREQ , SPI_CE_PIN); // initialize the LCD
-	myLCD.LCDFillScreen(0x8F); //splash screen bars
-	myLCD.setTextWrap(true);
-	bcm2835_delay(2500);
+bool setup() {
+	printf("LCD Test Begin\r\n");
+	// Check if Bcm28235 lib installed and print version.
+	if(!bcm2835_init())
+	{
+		printf("Error 1201: init bcm2835 library , Is it installed ?\r\n");
+		return false;
+	}else
+	{
+		printf("bcm2835 library Version Number :: %u\r\n",bcm2835_version());
+		bcm2835_delay(100);
+	}
+
+	if(!myLCD.LCDbegin(RAMaddressCtrl, LCDcontrast, SPICLK_FREQ , SPI_CE_PIN))  // initialize the LCD
+	{
+		printf("Error 1202: bcm2835_spi_begin :Cannot start spi, Running as root?\n");
+		return false;
+	}
+	printf("ERM19264 Library version number :: %u \r\n", myLCD.LCDLibVerNumGet());
+	myLCD.LCDFillScreen(0x33); // display splash screen bars, optional for effect
+	bcm2835_delay(1500);
+	return true;
 }
 
 void EndTest()
 {
 	myLCD.LCDPowerDown();
-	bcm2835_close(); // Close the library, 
+	bcm2835_close(); // Close library, deallocating allocated memory & closing /dev/mem
 	printf("LCD End\r\n");
 }
 
@@ -76,17 +87,23 @@ void TestReset()
 	myLCD.LCDclearBuffer();
 }
 
-void myTest()
+bool myTest()
 {
 
 	// define a buffer to cover whole screen 
 	uint8_t screenBuffer[myLCDwidth * (myLCDheight/8)]; 
 	myLCD.LCDbufferScreen = (uint8_t*) &screenBuffer;
+	if(myLCD.LCDbufferScreen== nullptr) // check if pointer is still = null
+	{
+		printf("Error 1203 :: Problem assigning buffer pointer\r\n");
+		return false;
+	}
 	myLCD.LCDclearBuffer();   // Clear buffer 
 	
 	// Call a function to display text
 	DisplayText();
 	DisplayGraphics();
+	return true;
 }
 
 // A series of tests to display the text mode